+++ pkg/peg/boot.go
@@ -33,24 +33,39 @@ func loadBootGrammar() (*Grammar, error) {
 		return nil, err
 	}
 	// these semantics apply only to parsers that parse grammars
-	g.Semantics = GrammarParserSemantics
+	g.Semantics = EBNFGrammarSemantics{}
 	return g, nil
 }
 
-// GrammarParserSemantics are the library's semantics for parsing grammars
+// EBNFGrammarSemantics are the library's semantics for parsing grammars
 // in the syntax defined by tatsu.ebnf and tatsu.json.
-func GrammarParserSemantics(
+type EBNFGrammarSemantics struct{}
+
+func (EBNFGrammarSemantics) Apply(
 	node trees.Tree,
 	ruleName string,
 	params []string) (trees.Tree, bool) {
-	if ruleName == "true" {
+	switch ruleName {
+	case "true":
 		return trees.TRUE, true
-	}
-	if ruleName == "false" {
+	case "false":
 		return trees.FALSE, true
-	}
-	if ruleName == "null" {
+	case "null":
 		return trees.NULL, true
+	case "meta":
+		text := textValue(node)
+		switch text {
+		case "name":
+			return &trees.Node{TypeName: "NameMeta"}, true
+		case "int":
+			return &trees.Node{TypeName: "IntMeta"}, true
+		case "uint":
+			return &trees.Node{TypeName: "UIntMeta"}, true
+		case "float":
+			return &trees.Node{TypeName: "FloatMeta"}, true
+		case "bool":
+			return &trees.Node{TypeName: "BoolMeta"}, true
+		}
 	}
 	// do nothing by default
 	// return false to tell caller to apply default processing